producer/requirements.txt
opencv-python>=4.8.0
httpx[http2]>=0.25.0
python-dotenv>=1.0.0

